refactor: (UI) Header 이미지버튼 컴포넌트 분리
